Correct information in the package.json
https://eaflood.atlassian.net/browse/WE-365

Spotted as part of building the initial application (after the first commit unfortunately) that the details in the `package.json` were incorrect and still referred to our trial project.

This change updates the details in `package.json` to match the project, and to include standard DEFRA/EA content.

It also renames the license file to match the DEFRA spelling convention.