Implement/integrate CI tools with the repo
https://eaflood.atlassian.net/browse/WE-388

We believe it is good practise to hook up our projects to some form of [Continuous Integration](https://www.thoughtworks.com/continuous-integration) (CI) as early as possible.

This change covers implementing and integrating a number of tools we use across the majority of our Node.js projects, the most critical of which is [Travis-CI](https://travis-ci.org) as this will automate running our unit tests and build scripts for every change to the repo.