Integrate code climate and travis-ci into project
Added a `.travis.yml` config file so we can control how Travis-CI builds the project.

Also integrated Codeclimate into the project. This will give us both code quality and visible test coverage metrics.

To do this means making changes to the Travis-CI config hence they are being done together.

One of the main changes with this commit is to the gulp file. We've added a CI version of the test task that outputs `LCOV` format files, as this is what Codeclimate needs to determine test coverage.