Rename crmToken service to activeDirectoryAuth
This renames `crmToken.service.js` to `activeDirectoryAuth.service.js` to reflect what the class is actually doing.

In this case we are working with the Azure Active Directory Authentication service to obtain a token, which we can then use when interacting with Dynamics in order to authenticate our requests.

So as its 'active directory authentication' which the class is responsible for, and in this case we are generating an `authentication token` this is the reason for the suggested renaming included in this change.