Refactor for consistency in config use
We have `/src/config/config.js` which is used to read and hold configuration values from the environment.

However we also had a couple of places where environment variables were read directly. This change aims to make consistent across the app where configuration values should come, and its the opinion of the committer that it should be `/src/config/config.js`.