Refactor setting host and path for https auth call
Prior to this change the details for the endpoint we need to connect to, and the token we need to pass in were held as a single env var.

However to make the call, we need to split the host i.e. domain from the path as part of setting the options for the `https.request` call.

Whilst we have control over our environment variables, it makes more sense to hold the values we need as required for the call, rather than performing any kind of transformation.

Therefore this change replaces the env var `CRM_TOKEN_ENDPOINT` with 2 new ones, and deletes the transformation logic in `/src/services/activeDirectoryAuth.service.js`.