Rename env vars prefixed crm to Dynamics
This change mainly renames the env vars prefixed with `CRM` to use `DYNAMICS` instead.

Early in the project we were advised to be refer to our backend system by its name (Dynamics) rather than its function (CRM). This change brings the project inline with this.

Also included are expanded comments and the removal of a env var that is only used if you have implemented <https://github.com/AzureAD/azure-activedirectory-library-for-nodejs>.