Remove hard coded config values from tests
Though the app gets it's config from environment variables, prior to this change the tests had the same values hard coded.

We expect these values to change, which would mean the tests would need to be updated. Also though not critical now (the credentials are for a temporary system with no actual data) we should always avoid exposing any credentials in our code.