Remove model validation from project
Experience on previous digital services has shown that attempting to validate models gets complex and error prone, due to the fact we generally collect a models values over multiple pages. Validating at the model level makes this hard, because for each page you have to have validation logic that understands both the current position of the transaction and the state of the model.

We moved away from this aproach in more recent services, and now use a pattern referred to as 'form objects' (though this might be specific to ruby). Essentially the validations for each form (or page) is managed in corresponding form object. Then essentially for each POST request, you can validate just within the context of the current form, before updating the corresponding model.