Remove TODO for uuid ver./Clean up other comments
HAving reviewed the details on the different versions of UUID (1 to 5) I'm happy that the version 4 spec is suitable for our needs. Namely that the risks of generating a duplicate ID are sufficiently small, and the fact it does not rely on a central authority (such as your mac address) gives us flexibility and possibly makes us more secure (I'm not sure the argument that it might be possible to identify the server a UUID was generated on is of concern to us).

So I have removed the TODO regards deciding on which version to use.

We also appeared to have TODO's regarding when we should be creating the session. So I have removed one of them along with some commented out code (IMHO something that never make it into your final commit).