Refactor contact 2 match Dynamics/fullname change
This actually makes 2 changes, but couldn't see how I could so them separately hence they are done as one.

The first is the suggestion we name the properties of our models the same as they appear in Dynamics. This reduces the amount of transformation logic we need to include, and makes it more explicit what Dynamics elements we are actually interacting with.

The second is the move from fullname to first name and last name. Have double checked the Defra strategy for data models, they have gone with holding users names as first and last separately (ugh!)

Having spoken with Martin and Adam, we are agreed that if this is the case we should leave it to the user to decide what to put in each field, rather than incorporating our own transformation logic. Hence the prototype is being updated to switch tio first name and last name, and the decision has been recorded in the decision log.

I have made the change here to move the project as close to current decisions as possible.