Refactor DynamicsDAL so GET requests have own func
Currently the `runQuery()` is handling to scenarios, retrieving data from Dynamics, and updating data in Dyanmics. This change (the first of 2) recommends we instead split it into 2 distinct functions.

This first change adds a new method specifically for handling `GET` requests to Dynamics.

N.B. The tests still pass, but I am now also getting an error I don't know how to resolve when I run them. Bit naughty but pushing for now as its more important the concept of the change is clear, not that the implementation is perfect.