Refactor DynamicsDAL for POST/PATCH requests
Previously the `runQuery()` was handling 2 scenarios, retrieving data from Dynamics, and updating data in Dyanmics. This change (the second of 2) recommends we instead split it into 2 distinct functions.

This change refactors the `runQuery()` method to just handle POST and PATCH requests, and also renames it `_commit()`.

N.B. The tests still pass, but I am now also getting an error I don't know how to resolve when I run them. Bit naughty but pushing for now as its more important the concept of the change is clear, not that the implementation is perfect.