Rename default to root route and controller
Following an initial review by the team it was felt there was some confusion between the default and base controller. So we instead decided to stick with the convention of naming controllers after their root.

Hence this change renames both the default controller and route to `root`. This should make it clearer going forward exactly which what is to be expected.