Rename DynamicsDAL to DynamicsDal
Follow feedback from the team, they accepted that referring to it as the Dynamics Data Access Layer helped, but that we should stick to hard camel case use. So even in the case where we are using an acronym it should still be camelcased.

Hence this change updates all references to the `/src/dynamicsDAL.service.js` to use `Dal` instead.