Replace capitilised element ID's
After some discussion amongst the dev team we have agreed to not use capitilised ID's for any elements in the views.

This change replaces an ID like `SITE_HEADING` with `site-heading`. We still need to confirm with the design team what their convention for ID's are, but for now this brings all elements inline.