Alan review of data-access-poc pull request (#8)
The purpose of this change is to connect our Node app to the Microsoft Dynamics instance and prove that we can perform a create, an update and read a list of Contacts from our application.

It also introduces an MVC folder layout for the project, with Models, Views and Controllers held in separate places.

Our unit tests have been expanded with the introduction of Sinon and Nock components. We also introduced the XMLDOM package to enable us to parse the HTML in responses in our unit tests and call methods such as getElementById() in order to test the content of our rendered templates as part of the unit tests.

We have also ensured we use ES6 classes everywhere and that our controllers and models both have a base class in which we can implement common functionality if required.

Note: At this point we have not carried out a full unit test of the HTML content in our templates (e.g. Site, Contacts, etc.) as this was not in scope for this PR. The unit testing of the HTML templates that is done is intended to be illustrative of the methods to be used, as it is anticipated that these templates will have changed by the time we come to implement them for real.

Also some functionality e.g. `Contact.getById()` has been mocked as the key thing we were attempting to do was prove we could do basic CRUD with Dynamics, without actually implementing functionality.