Implement getById() / refactor Contact.list()
This change properly implements the `Contact.getById()` function so that now it will take an ID, query Dynamics, and return an instance of `Contact`.

To do this and reuse `DynamicsDalService._query()` I've also had to refactor what it returns. This is because when you don't specify an ID you'll get `{ "value": [] }`. However when you query by ID the result is just the object e.g. `{ "firstname": "Foo", ... }`.

So now the `_query()` method simply returns the result, and its up to the calling method to transform it to the result required.

Finally I've also amended `Contact.list()` to return an object that holds the results as an array, plus a count e.g. { "count": 1, "results": [ { "firstName": "Foo", ... } ] }`.

Currently to see `getById()` in action you need to enter an ID at `/search` and then check the console. Room for improvement admittedly!