Re-implement handle no authToken in contacts ctrlr
When working to get the update working I found that the `contact.route.test.js` was broken. To test a route it appears we need to rely on Hapi's server.inject method. The problem is it doesn't support passing a session object.

Therefore it looks like to support our tests, we can't assume that there will always be a session when it comes to the controller endpoints because this will break the tests.

This change puts back in some defensive code for there not being a session.