Fix typos / rename 'contacts' to 'contactSearch'
Following feedback on the PR this change fixes some typos. The main change though is renaming files related to the `contacts` resources as `contactSearch`.

Because we had to have it on the path `/search` to avoid conflicts with the existing `contact` resource, it was noted this was confusing when everything else (route, controller and view) referred to `contacts`. Therefore this change updates the resource to be called `contactSearch` and the path is now `/contact-search`.