Fix not setting config values breaks tests in CI (#13)
* Fix not setting config values breaks tests in CI

We recently spotted that our tests were actually failing on our CI server, yet the build was still being reported as good. PR #12 fixed the problem of the build not failing, this change resolves the problem causing the tests to fail.

Basically because of changes made to how we do config, and ensuring we always use config rather than having hard coded values if the relevant env vars don't exist, the tests will fail.

This change resolves the problem by ensuring we have safe i.e. ok to commit default values for when an env var is missing. In environments like our CI server, it means we don't have to set env vars but the tests will still be able to run as expected.