Working example of front end validation
https://eaflood.atlassian.net/browse/WE-587

The purpose of this PR is to implement a working example of front end validation in our application using Joi, our chosen validation framework:

https://github.com/hapijs/joi

Validation has been added to our example 'Site' page. To make the validation more realistic I added a new 'Another Name' field to the page. This allows us to demonstrate having more than one error on the page.

A new Handlebars helper (getFormGroupClass.js) was added that allows a class in the HTML to be conditionally set based on whether or not the corresponding field contains a validation error.

This PR also includes the following improvements:

* Page heading and title refactoring

* Refactoring of our route paths and pate headings into the Constants file

* Refactored creation of page context and added a Handlebars helper to get the error class, simplifying the HTML in the templates.

* Added the ‘Problem: ‘ prefix to the page title if there is an error.
Refactored the controllers to avoid repeated code.

* Disabled HTML linting on IDs and classes to avoid Handlebars HTML validation issue.