Implement the Contact.getById() function (#9)
Implement the Contact.getById() function

Pior to this change the `Contact.getById()` mocked a return value. This is fine for getting us up and running but we now need to implement the actual functionality so that we move towards all code and features in the project being 'working' code.

This change replaces the mock result with an actual result, by implementing a call to Dynamics to return a result given a specified ID, and then transforming that into an instance of `Contact.js`.

It also adds a new page (route, controller and view) to demonstrate the functionality, along with moving the existing `Contact.list()` example to it as well.