Update dynamicsDal.service.test.js to match change
Tests updated to reflect the changes to `DynamicsDal.service._commit()`. This is called via the public `createItem()` and `updateItem()` methods, both of which will return the ID of the entity being saved.

The difference is that for `createItem()` the ID will have been generated, and for `updateItem()` it should match the ID of the entity passed in.