Drop 'Item' from createItem() & updateItem in DAL
Suggested name change for 2 of the public methods in `DynamicsDalService`. Specifically I don't think we need to postfix the word 'Item' on the `create()` and `update()` methods. What we are creating or updating is defined by the dataObject argument passed in. I also think it starts to build a simple, clean API

- create()
- update()
- search()