Gulp task to check for {{{ in Handlebars code
https://eaflood.atlassian.net/browse/WE-625

The purpose of this change is to create a Gulp task to be executed as part of our normal build / test process that will check for occurrences of {{{ in our HTML views. If any are found the Gulp task will report the problem and exit.

The Gulp task is also added to the Travis CI build and if any {{{ are detected in our HTML views then we ensure that the Travis CI build fails.