Add method 2 BaseModel 2 determine instance is new
It looks like for the majority of our models we are going to have to determine if a PATCH request is needed for an update, or a POST for a create.

Currently we determine that by checking if the instance has an ID. Rather than stick `if (this.id)` throughout our different instances a good idea was to create a new method on the base class called `isNew()`. All instances and external classes could use this method to determine if the instance exists in Dynamics or not, and should we change how we make the determination we only have to make the change in one place.