Refactor DynamicsDalService 2 remove duplication
This started with an effort to move the logic for extracting the ID from a Dynamics 204 response into its own method. In that way a big block of text explaining what its doing would fit better (rather than slap in the middle of a method).

However to do that I needed to refactor `_commit()` to simply return the response it called, and get the calling method (`create()` or `update()`) to then call the new `_extract()` method.

Well having simplified `_commit()` to be 'call Dynamics and return result' it smacked of being overly similiar to `_query()`. So also included in this change is a refactor of `DynamicsDalService` to merge `_query()` and `_commit()` back into one method. What is specific to `create()`, `update()` and `query()` is handled by those methods, and we have a new `_setCallSpecificRequestOptions()` to handle setting the request options.