Imp. extract and setting of ID after contact save
Imp. extract and setting of ID after contact save

Currently the project has example code of creating a new contact record in Dynamics. What it doesn't have is how to extract the ID from what Dynamics returns, and then set that against our instance of the contact.

This change implements the functionality, so we have it as an example to work with going forward.

It also focuses on refactoring `DynamicsDalService` to remove some of the duplication, and adds a new method to `base.model.js` to allow us to have a consistent way of checking if an entity is new or has been persisted to Dynamics at some point.