Move gulp build dep's from dev to main
We rely on gulp to build the app, but it was currently listed as a dev dependency. It along with those used to convert the sass to css need to be available when running in production mode so that we can run `gulp build` without relying on any global dependencies.