Move more dependencies from dev to prod
At the moment just trying to get past issues with missing dependencies when running `gulp build` in a production environment.

A number of this dependencies only relate to checks that are done before the project would be deployed, so will look later to see how we can move them back out.