Refactor how we obtain current git commit sha
Currently we have a gulp task that writes to file (`latestCommit.json`) what the sha is for the latest git commit present in the app.

However when the app is deployed to Heroku this method no longer works because

- git information is not available to a running app
- an app in Heroku cannot write to file

The good news is Heroku does populate an environment variable with the git sha during the build process (see https://devcenter.heroku.com/articles/buildpack-api#bin-compile-summary). The bad news is this env var is also not available to running apps. However a buildpack has been created that can extract the value from `SOURCE_VERSION` and make it available to a running app as `GIT_SHA`.

We can add the buildpack to our app by

- installing the [Heroku CLI](https://devcenter.heroku.com/articles/heroku-cli)
- logging into heroku from the cli with `heroku login`
- adding the build pack to the app with `heroku buildpacks:set https://github.com/dive-networks/heroku-buildpack-git-sha -a waste-permits`

We could have then changed the app to also check for this if looking for the file failed. However instead this change tweaks how we get and set the git sha. Rather than create a file now the gulp task sets an environment variable instead. As an environment variable we can then read it with `Config`. When you run the gulp task it will create an env var called `GIT_SHA`. When you run the app it will look for this env var irrespective of who set it.