Refactor how we obtain current git commit sha (#31)
Previously we had a gulp task that wrote to a file (`latestCommit.json`) what the sha is for the latest git commit present.

However when the app is deployed to Heroku this method no longer works because

- git information is not available to a running app
- an app in Heroku cannot write to file

The good news is Heroku does populate an environment variable with the git sha during the build process (see https://devcenter.heroku.com/articles/buildpack-api#bin-compile-summary). The bad news is this env var is also not available to running apps. However a buildpack has been created that can extract the value from `SOURCE_VERSION` and make it available to a running app as `GIT_SHA`.

We can add the buildpack to our app by

- installing the [Heroku CLI](https://devcenter.heroku.com/articles/heroku-cli)
- logging into heroku from the CLI with `heroku login`
- adding the build pack to the app with `heroku buildpacks:add https://github.com/dive-networks/heroku-buildpack-git-sha -a waste-permits`

We could have then changed the app to also check for this if looking for the file failed. However instead this change tweaks how we get and set the git sha. Rather than create a file, now the gulp task sets an environment variable instead. As an environment variable we can then read it with `config.js`. When you run the app locally using the default gulp task the env var `GIT_SHA` will be created. When you run the app in Heroku it will already exist, and either way `config.js` will pick it up.