Amend to support deploy and run app in Heroku
As a means of checking that the app can deploy and run in a production like environment whilst we await our actual environments, we would like to deploy the app to [Heroku](https://www.heroku.com/).

However to support this we need to make a number of minor changes to the project

- Change the name of the env var we use to get the port number from. On Heroku you cannot set the port; it is dynamically applied. So your app must read it from an env var named PORT. Rather than tweak the code to do 2 checks (check for our env var, then check for Heroku's) this change amends the code to always just look for an env var name PORT
- Update the Procfile to follow Heroku best practise for a node app running on a free dyno
- Add the `postinstall` directive to `package.json` so that we run our gulp build tasks after an npm install. This is something Heroku expects

Additionally in Heroku we have added all the env vars we need (excluding PORT as Heroku sets that), along with setting `NPM_CONFIG_PRODUCTION` to false. This is because our gulp file contains tasks that are only run in development, but needs has to require dev only dependencies to make that happen. By default Heroku will install the dependencies using `npm install --production` which would ignore anything under `devDependencies`. Setting this env var ensures everything gets installed.