Apply for a Standard Rules Rules waste permit
https://eaflood.atlassian.net/browse/WE-636

This PR adds the "Apply for a standard rules waste permit" page. This is the starting point for the workflow so a cookie is created and added to the request when the Continue button is pressed. A call to Dynamics creates a new Application and the corresponding ID is added to the cookie.

A new CookieService has been created which is used to generate and validate cookies.

This PR also includes the initial versions of the follow-on routes:

/permit-category
/save-and-return/check-your-email