Add missing skipLinkMessage to context
https://eaflood.atlassian.net/browse/WE-636

It was spotted when testing the accessibility of the 'APply for a standard rules permit' (the first page in the service to be properly implemented) that the option to skip to the main content was not working properly.

Upon investigation we believe its simply because we are not providing the text for the link, hence nothing is displayed. And to do that we need to add `skipLinkMessage` to the context object.

This change fixes the issue.