Sign session cookie with password
In order to improve security this change amends the config we provide Hapi to create the session cookie, to tell it the cookie must be [signed](https://hapijs.com/api#serverstatename-options).

Specifically we will provide Hapi with an object that contains a password that will be used to calculate an [HMAC](https://en.wikipedia.org/wiki/Hash-based_message_authentication_code) for cookie integrity validation. This does not provide privacy, but a means to verify that the cookie value was generated by the server.