Refactor the logging service
The purpose of this PR is to refactor the original Server Logging Service so that a single service can now carry out Request and Server logging.

An optional 'request' parameter is passed into the logging method. If present then the log entry goes to the request object, otherwise the server object logs the item.

Given that both types of log are now handled by the same service, the service has been renamed to the LoggingService.