Fix missing commit hash if app started directly
If the service is run on Heroku than we have no issue with the commit hash. We have added a buildpack that reads the latest git commit during build, and then creates an env var called `GIT_SHA` with the hash as its value.

Locally if you start the app using gulp it also resolves the issue. There is a gulp task which is called when the default task is run, and it uses a gulp addon to obtain the value, and then also creates an env var called `GIT_SHA`.

The problem is if you don't use gulp to start the app, for example `node server.js` or as it came to light for me `pm2 start server.js`, you never get an env var called `GIT_SHA` created.

This means on the version and healthcheck page the commit hash displays as `unknown`. This fix is to implement a solution for when the app is not called from either Heroku or gulp.