Add missing IDs to form fields
If a validation error message is displayed, users should be able to click on the error message and be redirected to the form field where the problem is. However, if the form fields (or surrounding elements) don't have the correct ID, this doesn't work. This change adds the correct IDs to all form fields so the error message validation links work properly.