Sign and rename the crumb crsf cookie
https://eaflood.atlassian.net/browse/WE-619

During QA of the original implementation of [Crumb](https://github.com/hapijs/crumb) we spotted that

- It creates a cookie called crumb
- The value in the cookie matches the value in the hidden input field in a pages form

Manipulating it with Chrome dev tools meant as long as the 2 values matched, the token would be accepted. Initially this made us suspect if this implementation of CSRF tokens was valid, but upon review we realised this was possible because it was 'my' session within the browser we were manipulating. The token is to prevent others exploiting my session and because we will be tranferring the data via https this will not be possible.

We have agreed to make 2 changes though to our existing implementation to improve it slightly

- We will rename the cookie from crumb to something more meaningful to provide confidence to the user as to its source
- We will sign the cookie as this should prevent anyone from tampering with the cookie value