Set options on crumb cookie 2 match session cookie
This change applies the same settings we assign to the main session cookie to the one Crumb creates for managing the CSRF token.

This also means the cookie is signed. Now when viewing in the browser using dev tools you will see that the token in the form and the one in the cookie have different values. Also if you amend the `DefraCsrfToken` cookie you'll get an invalid cookie error. And if you amend the token in the form you'll get a 403 forbidden error. Attempting to set the two tokens to the same value will also result in an invalid cookie error.