Start to remove custom proxy support
Initially ruled out using something like [node-https-proxy-agent](https://github.com/TooTallNate/node-https-proxy-agent) on account that it had no license. However the entry on NPM records it as MIT, and double checking the package.json file it also confirms its MIT.

So rather than trying to build up the headers ourselves to support using a proxy, now focusing on implementing this package within the app.

This change covers removing some of the branching `if proxy` logic, but retaining the setting of the request options in the separate method.