Set a timeout for http requests from the app
https://eaflood.atlassian.net/browse/WE-416

Whilst trying to get the service running within our development environment we found were getting timeouts. Identified the issue as the need to support using a proxy server (see PR #53) but it did highlight that if we loose connection it takes a very long time for the attempt to connect to finally timeout.

Hence this change which will implement the ability to timeout an attempt much sooner based on a configurable value.