Refactor dynamics srv. to base options on full url
Looking into why proxy support seems to be working for the active directory authentication service but not the call to Dynamics (keep getting a 401 not authorise response).

On tha basis that we are not actually setting up the request `options` object properly, this refactors the service so that when we initialise the object we do so passing in the full url to `url.parse()`. Previously we would be overwriting the path and specifying the method based on the call what kind of call was made e.g. search, update, and the query passed into it.

Also a bunch-o `console.log` calls!