Implement proxy server support for external calls (#53)
* Implement proxy server support for external calls

https://eaflood.atlassian.net/browse/WE-416

Having recently deployed our app to our development environment, we've been reminded that when the app needs to make a connection to the internet (for example when communicating with Dynamics) it will have to make that connection through a proxy server.

So this change updates our services that make http requests to external services to support using a proxy. The change also handles when working locally that we don't have a proxy server running.