Change name of GIT_SHA file
https://eaflood.atlassian.net/browse/WE-416

When the app is deployed there will be no `.git` folder so we can't query git for the latest commit sha, nor will there be the env var GIT_SHA.

We are intending to use [Shipit](https://github.com/shipitjs/shipit) for our deployment, and initially we thought the only way to obtain the commit sha was to add a custom action that created a file called GIT_SHA containing with the sha.

However after more use of the tool, we spotted it already does this automatically, but calls the file `REVISION`. So we've been able to strip out the custom deployment action, however it does mean a change in the name of the file we expect when the env var and querying git is not available.