Task list (#58)
https://eaflood.atlassian.net/browse/WE-563

This change introduces the Task List page to the application.

The Task List is built up using a hard-coded list of sections and section items contained in the model. Dynamics is queried by passing in the permit code (e.g. 'SR2015 No 18') and the IDs of the task list items of interest and Dynamics returns a set of Booleans to tell us whether or not the task list item should be available (visible) in the Task List.

The intention is to hard code the sections and task list items in the StandardRule model in order to keep the HTML simple and to avoid repetition in the HTML, using a repeating HTML task list template in a nested for...each loop.

It is assumed that all Task List items are turned off by default. We turn on the ones that Dynamics tells us should be available for the particular permit type that has been selected.

The final Task List item ('Send application and pay') isn't really a task list item like the others. It is really a submit button in disguise. It won't have a completion status - all it does is act as a call to action, performing validation and forwarding to the 'check your answers' page if validation succeeds. Hence the 'Send application and pay' task list item isn't held in Dynamics, so we have to code around that to some extent to ensure it is always available and without breaking the model and view.