Page back links
The purpose of this change is to add Back links to all relevant pages and to create corresponding unit tests.

The Back link is built using Javascript therefore it is hidden if the user does not have Javascript available. It was decided that the Back link is a progressive enhancement since a non-Javascript user can press the browser back button instead if the Back link is not available.