Add current path to the health page
https://eaflood.atlassian.net/browse/WE-416

In our main environments whenever the app is deployed a new folder is created, named using the timestamp of the current date and time. The curent health page will display a commit hash of the version of the project we are running but the same version of the app could be deployed multiple times.

So for debugging purposes it would be useful to see specifically which copy of the project is running, and we can determine that by knowing which release folder it is currently running out off. This change adds the current path to the health page to allow us to see this.