Fixed site name length
https://eaflood.atlassian.net/browse/WE-470

This change fixes a bug that was identified regarding the length of the Site Name text field on the Give Site Name page. The Site Name text field was shorter than expected.

The fix is to include some additional SCSS into our application SCSS file. This had been added to the Waste Permits prototype but it hadn't been added to the Waste Permits application.