Fixed test problems following upgrade to Lab version 15
Lab version 15 drops support for ending tests by executing the done callback and running tests in parallel. All tests now are executed using async/await and should be much simpler to construct as a result. As a result, it was necessary to remove all the calls to done() from our tests.