Fix unit tests so that they work in v15 of Lab
https://github.com/hapijs/lab/issues/770

Lab version 15 drops support for ending tests by executing the done callback and running tests in parallel. All tests now are executed using async/await and as a result, support for Node.js versions prior to 8 are no longer supported.

This was a breaking change that required all of our unit tests to be amended to use async/await rather than callbacks.