Add DEBUG logging to the application
The purpose of this change is to add DEBUG logging capability to our application. The log level is now set in the .env file.

If LOG_LEVEL=INFO then there is no change in behaviour, the ERROR and INFO level logs are created as before.

If however LOG_LEVEL=DEBUG then all debug messages in the code will be created.

This will be useful when debugging issues such as, for example, the retrieval of data from Dynamics.

If debugging is turned on with LOG_LEVEL=DEBUG then we also now put Hapi server in debug mode. This configures the server to log a detailed stack trace to the console for any routes that return an internal server error.