Give site grid reference page
https://eaflood.atlassian.net/browse/WE-472

The purpose of this change is to add the 'Give site grid reference' page to the application.

The page allows the user to enter the a 12 digit grid reference to specify the location of the Site. The site grid reference is held in a Dynamics LocationDetail object, as opposed to the site name which is held in the parent Location object in Dynamics, therefore this change required a fair amount of refactoring to ensure that the completeness check continues to function and the app will scale as we add more and more controllers and model classes.