Enter site postcode and retrieve address list
https://eaflood.atlassian.net/browse/WE-473

This change adds the Enter Postcode screen along with the Select Address screen. When the postcode is entered, the AddressBase facade is queried and a list of addresses is returned that is used to populate the drop-down list in the Select Address screen.

Note that this development is "happy path" only for now so there is very little in the way of error handling (e.g. handling invalid postcodes or returning zero addresses for a valid postcode).

This story is to be revisited in future sprints and is no longer considered to be part of our initial Minimum Viable Product.