Confirm correct company has been selected
https://eaflood.atlassian.net/browse/WE-591

This adds the initial version of the 'Confirm correct company has been selected' page. The work is not complete as it depends on work that is being carried out on the previous page in the flow (enter company name), but it is difficult to progress this change any further at the moment. This page assumes that an Account entity has already been created in Dynamics, which is created by the previous page in the flow.

This change includes the initial version of the lookup service that obtains data from Companies House, which will be required by another of other pages in the Company Details page flow.

Some refactoring of the unit tests is also included, such as:

- Standardising the ID of the H1 page headings to be 'page-heading' and the ID of the page call to action buttons to be 'submit-button'
- Testing to ensure that the defra-csrf-token element is included in the page
- Using a checkPageElements function in the tests that loops through an array of page element IDs to check if they exist