Site postcode screen - fixed address lookup
This is just a small change to fix the address lookup on the site postcode screen. We parked this previously but but the postcode screen fails when the page is submitted without this fix.

Also it will be useful to be able to retrieve addresses from the addressbase facade using the postcode screen.

We are likely to be going directly to Dynamics for addresses in future but this change allows us to see the Postcode and Select Address screens when we carry on with their development.